render: the line clipper cut a line one dot further on one machine
The reference frame that differed between the two architectures was a fused
multiply-add, and reading for it found three sites that were not the one.

The search was made to answer instead of guessing. The prepared shape hashed
the same on both machines, and so did the fitted view. Switching each
basemap layer off in turn showed that water alone carried the difference,
and the differing cells were a diagonal at the frame's left edge - which is
where a line is cut to the rectangle.

Then the compiler was asked directly. Building with fused multiply-adds
turned off made this machine draw exactly what the other draws, which proved
the class of fault; bisecting on the compiler's own hash of each fused site
named the renderer and then the clipper: four fusions in one line, at the
moment a line meeting the edge is cut and rounded to a whole dot. A machine
that fuses keeps more bits and cuts the line one dot further along.

Every product there is converted before it is added now, and the two
architectures draw the same frame. Three more sites are hardened on the way
and are worth keeping: the simplifier's distance, the view's origin, and a
tile's corner - each the same latent fault waiting for a different view.

+++ internal/project/view.go
@@ -71,7 +71,13 @@ func (v View) origin() (x, y float64, err error) {
 		return 0, 0, err
 	}
 	w := TileSize * math.Exp2(v.Zoom) // the side of the whole world, in dots, at this zoom
-	return cx*w - float64(v.Cols*DotsPerCol)/2, cy*w - float64(v.Rows*DotsPerRow)/2, nil
+	// **Each product is converted before the half-window is taken off it**,
+	// so that no machine fuses the two into one rounding. This is the
+	// origin every dot of a frame is measured from: a fused multiply-add
+	// keeps more bits than the separate steps, so a machine that fuses
+	// puts a line's dots in different cells from one that does not, and
+	// the same data draws two pictures (NFR-6, constants section 6).
+	return float64(cx*w) - float64(v.Cols*DotsPerCol)/2, float64(cy*w) - float64(v.Rows*DotsPerRow)/2, nil
 }
 
 // ToDot projects a position into the view's dots: x grows east and y grows
@@ -146,5 +152,11 @@ func (v View) TilePlace(t scene.TileID) (x, y, side float64, err error) {
 		return 0, 0, 0, err
 	}
 	side = TileSize * math.Exp2(v.Zoom-float64(t.Z))
-	return float64(t.X)*side - ox, float64(t.Y)*side - oy, side, nil
+	// **The product is converted before the origin is taken off it**, so
+	// that no machine fuses the two into one rounding: a fused
+	// multiply-add keeps more bits than the separate steps, and this is
+	// where a tile's corner is decided. A corner a fraction of a dot apart
+	// puts a line's dots in different cells, and the same data then draws
+	// two pictures on two machines (NFR-6, constants section 6).
+	return float64(float64(t.X)*side) - ox, float64(float64(t.Y)*side) - oy, side, nil
 }
